+++ holidays/groups/islamic.py
@@ -14,7 +14,6 @@ from collections.abc import Iterable
 from datetime import date
 
 from holidays.calendars import _IslamicLunar
-from holidays.calendars.gregorian import _timedelta
 from holidays.groups.eastern import EasternCalendarHolidays
 
 
@@ -442,20 +441,3 @@ class IslamicHolidays(EasternCalendarHolidays):
         return self._add_islamic_calendar_holiday(
             name, self._islamic_calendar.tasua_dates(self._year)
         )
-
-    def _is_long_ramadan(self, eid_al_fitr: date) -> bool:
-        """Check whether the Ramadan preceding the given Eid al-Fitr date lasted 30 days.
-
-        Args:
-            eid_al_fitr:
-                The date of Eid al-Fitr.
-
-        Returns:
-            True if Ramadan lasted 30 days, False otherwise.
-        """
-        first_ramadan_dates = {
-            d[0]
-            for y in (self._year - 1, self._year)
-            for d in self._islamic_calendar.ramadan_beginning_dates(y)
-        }
-        return _timedelta(eid_al_fitr, -30) in first_ramadan_dates
